test(worker-sessions): keep projection regression within file limits
